[migration + file browse][new migration for postgresql, on browse image preview, product remove, error stripe to respective tab]

+++ inzaana_cms/app/Http/Controllers/ProductController.php
@@ -4,12 +4,15 @@ namespace Inzaana\Http\Controllers;
 
 use DB;
 use Auth;
+use Log;
 use Validator;
+use Exception;
 use Illuminate\Http\Request as ProductRequest;
 
 use Inzaana\Http\Requests;
 use Inzaana\Http\Controllers\Controller;
 use Inzaana\Product;
+use Inzaana\Store;
 use Inzaana\MarketProduct;
 use Inzaana\Category;
 use Inzaana\Mailers\AppMailer;
@@ -17,6 +20,9 @@ use Inzaana\BulkExportImport\ProductImporter;
 
 class ProductController extends Controller
 {
+    const PRODUCT_ENTRY_TABS = [ 'single_product_entry_tab', 'bulk_product_entry_tab' ];
+
+
     /**
      * Create a new controller instance.
      *
@@ -39,7 +45,20 @@ class ProductController extends Controller
         $productsCount = 0;
         $products = Auth::user()->products;
         $categories = Category::all();
-        return view('add-product', compact('productsCount', 'products', 'categories'))->with('user', Auth::user());
+        $tab = !session()->has('selected_tab') ? self::PRODUCT_ENTRY_TABS[0] : session('selected_tab');
+        $viewData = [ 
+            'productsCount' => session()->has('productsBySearch') ? session('productsBySearch')->count() : 0,
+            'productsBySearch' => session()->has('productsBySearch') ? session('productsBySearch') : [],
+            'products' => Auth::user()->products,
+            'product' => session()->has('product') ? session('product') : null,
+            'categories' => Category::all(),
+        ];
+        session()->forget('selected_tab');
+        session()->forget('productsBySearch');
+        session()->forget('product');
+        return view('add-product', $viewData)->withUser(Auth::user())
+                                             ->withStores(Auth::user()->stores->pluck('name', 'name_as_url'))
+                                             ->withTab($tab);
     }
 
     /**
@@ -130,74 +149,71 @@ class ProductController extends Controller
             $search_terms = $request->query('search-box');
             $search_terms_slugged = str_slug($search_terms);
 
-            $productsBySearch = Product::where('title', $search_terms)->orWhere('title', 'like', '%' . $search_terms . '%')->get();;
+            // NOTE: like -> searches case sensitive
+            // $productsBySearch = Product::where('title', $search_terms)->orWhere('title', 'like', '%' . $search_terms . '%')->get();
+
+            // NOTE: ilike -> seaches case insensitive
+            $productsBySearch = Product::where('title', $search_terms)->orWhere('title', 'ilike', '%' . $search_terms . '%')->get();
             $productsCount = $productsBySearch->count();
-            return redirect()->route('user::products.search-terms', [$search_terms_slugged])
-                                ->with(compact('productsBySearch', 'productsCount'));
+
+            return redirect()->route('user::products')
+                             ->with(compact('productsBySearch', 'productsCount'));
         }
-        return redirect()->route('user::products')->with(compact('products', 'categories'));
+        return redirect()->route('user::products');
     }
 
-    public function searchTerms($terms)
-    {   
-        $productsCount = session()->has('productsCount') ? session('productsCount') : 0;
-        $categories = Category::all();
-        $products = Auth::user()->products;
-        $productsBySearch = session('productsBySearch');
-        return view('add-product', compact('productsCount', 'products', 'categories', 'productsBySearch'))->with('user', Auth::user());
-    }
-
-    public function edit($product_id)
+    public function edit(Product $product)
     {
         //edit the product
-        $productsCount = 0;
-        $products = Auth::user()->products;
-        $product = $products->find($product_id);
-        $categories = Category::all();
-        return view('add-product', compact('productsCount', 'products', 'categories'))->withUser(Auth::user());
+        return redirect()->route('user::products')->withProduct($product);
     }
 
-    public function delete($product_id)
+    public function update(ProductRequest $request, Product $product)
+    {
+        return response($product, 200);
+    }
+
+    public function delete(Product $product)
     {
         //delete the product
-        $products = Auth::user()->products;
-        $product = $products->find($product_id);
-        $message = 'Your product (' . $product->title . ') is removed from your product list.';
-        if($product)
+        if($product)// && $product->marketProduct() && $product->marketProduct()->delete()
         {
+            $message = 'Your product (' . $product->title . ') is removed from your product list.';
             $product->delete();
+            flash($message);
         }
         else
         {
             $message = 'Your product (' . $product->title . ') is NOT removed from your product list.';
             $message .= 'Product is NOT found to your product list.';
             $message .= 'Contact your administrator to know about product removing policy';
+            flash($message);
         }
-        flash($message);
         return redirect()->route('user::products');
     }
 
-    public function copy($id)
+    public function copy(Product $product)
     {
         // $products = Auth::user()->products;
-        $product = Auth::user()->products()->find($id);
+        $productExists = Auth::user()->products()->find($product->id);
         // $product = Product::find($id); 
-        if($product)
+        if($productExists)
         {
-            flash()->warning('Your selected product (' . $product->title . ') is from YOUR product list.');
+            flash()->warning('Your selected product (' . $productExists->title . ') is from YOUR product list.');
         }
         else
         {
-            $product = Product::find($id);
             if(!$product)
             {
                 return redirect()->back()->withErrors(['Your selected product is invalid. Please conatct your administrator.']);
             }
             $product->user_id = Auth::user()->id;
             $productSerialized = collect($product->toArray());
+
             $productSerialized->forget('id');
             $productSerialized->forget('created_at');
             $productSerialized->forget('updated_at');
+
             $productCopied = Product::create($productSerialized->toArray());
             if(!$productCopied)
             {
@@ -256,11 +272,100 @@ class ProductController extends Controller
         return redirect()->back();
     }
 
+    private function getUploadedFileName(ProductRequest $request)
+    {
+        $errors = [];
+        if (!$request->hasFile('csv'))
+        {
+            $errors['has_input'] = 'No file is uploaded';
+            flash()->error($errors['has_input']);
+            return $errors;
+        }
+        $uploadingFile = $request->file('csv');
+        if (!$uploadingFile->isValid())
+        {
+            $errors['corrupted_file'] = 'Upladed file is corrupted';
+            flash()->error($errors['corrupted_file']);
+            return $errors;
+        }
+
+        if($uploadingFile->getSize() >= $uploadingFile->getMaxFilesize())
+        {
+            $errors['size_limit_exits'] = 'Please keep your file size below ' . ($uploadingFile->getMaxFilesize() / 1000);
+            flash()->warning($errors['size_limit_exits']);
+            return $errors;
+        }
+        if(!ProductImporter::isSupportedExtension($uploadingFile->getClientOriginalExtension()))
+        {
+            $errors['ext_not_supported'] = 'Upladed file (*.' . $uploadingFile->getClientOriginalExtension() . ') is not supported!';
+            flash()->warning($errors['ext_not_supported']);
+            return $errors;
+        }
+
+        //Display File Name
+        $out = 'File Name: '.$uploadingFile->getClientOriginalName();
+        Log::info('[Inzaana][' . $out . ']');
+        //Display File Extension
+        $out = ', File Extension: '. $uploadingFile->getClientOriginalExtension();
+        Log::info('[Inzaana][' . $out . ']');
+        //Display File Real Path
+        $out = ', File Real Path: '.$uploadingFile->getRealPath();
+        Log::info('[Inzaana][' . $out . ']');
+        //Display File Size
+        $out = ', File Size: '.$uploadingFile->getSize();
+        Log::info('[Inzaana][' . $out . ']');
+        //Display File Mime Type
+        $out = ', File Mime Type: '.$uploadingFile->getMimeType();        
+        Log::info('[Inzaana][' . $out . ']');
+        try
+        {
+            //Move Uploaded File
+            $serverFile = $uploadingFile->move(ProductImporter::getStoragePath(), $uploadingFile->getClientOriginalName());
+            if(!$serverFile)
+            {
+                $errors['file_move_error'] = 'Upladed file did not move!';
+                flash()->error('Something went wrong during upload. We have already logged the problems. Please contact Inzaana help line for further assistance.');
+                return $errors;
+            }      
+            Log::info('[Inzaana][ Upload success to ' . $serverFile->getRealPath() . ']');
+            return $serverFile->getBasename();
+        }
+        catch(\Symfony\Component\HttpFoundation\File\Exception\FileException $fe)
+        {
+            $errors['unknown_file_error'] = '[Inzaana][ Upload failed: ' . $fe->getMessage() . ']';
+            Log::critical($errors['unknown_file_error']);
+        }
+        return $errors;
+    }
+
     public function uploadCSV(ProductRequest $request)
     {        
         try
         {
-            $pi = new ProductImporter('product_inzaana_asset.csv');
+            $result = $this->getUploadedFileName($request);
+            if(is_array($result) && count($result) > 0)
+            {
+                session(['selected_tab' => self::PRODUCT_ENTRY_TABS[1]]);
+                return redirect()->back()->withErrors($result);
+            }
+
+            if(!$request->has('stores'))
+            {
+                session(['selected_tab' => self::PRODUCT_ENTRY_TABS[1]]);
+                return redirect()->back()
+                                 ->withErrors(['Please give a store name to upload your products.']);
+            }     
+                
+            $store_name_as_url = $request->input('stores');
+            $store = Store::whereNameAsUrl($store_name_as_url)->get()->first();
+            if(!$store)
+            {
+                session(['selected_tab' => self::PRODUCT_ENTRY_TABS[1]]);
+                return redirect()->back()
+                                 ->withErrors(['We did not find any store name of yours named ' . $store_name_as_url . '.inzaana.com.']);
+            }
+
+            $pi = new ProductImporter($result);
             $csv = $pi->getProducts()['raw'];
             foreach($csv as $value)
             {
@@ -276,11 +381,11 @@ class ProductController extends Controller
 
                 if(!$marketProduct->save())
                 {
-                    return 'Market failed';
+                    Log::info('[Inzaana][User:: ' . Auth::user() . '][error:: Market product (' . $value['title'] . ') not added.]');
                 }
                 $product = new Product();
                 $product->user_id = Auth::user()->id;
-                $product->store_id = $store_id;
+                $product->store_id = $store->id;
                 $product->market_product_id = $marketProduct->id;
                 $product->is_public = $value['is_public'];
                 $product->title = $marketProduct->title;
@@ -292,17 +397,20 @@ class ProductController extends Controller
                 if(!$product->save())
                 {
                     $marketProduct->forceDelete();
-                    return 'Product NOT uploaded';
+                    Log::info('[Inzaana][User:: ' . Auth::user() . '][error:: Product (' . $value['title'] . ') not added.]');
                 }
                 if(!$product->saveDiscountedPrice())
                 {
-                    return 'MRP not saved';
+                    //return 'MRP not saved';
+                    Log::info('[Inzaana][User:: ' . Auth::user() . '][error:: Product (' . $value['title'] . ') MRP not saved.]');
                 }
             }
-            return Product::where('special_specs->camera->values', 10)->get();
+            flash()->success('Successfully uploaded all products.');
+            return redirect()->back();//Product::where('special_specs->camera->values', 10)->get();
         }
         catch(\Exception $e)
         {
+            session(['selected_tab' => self::PRODUCT_ENTRY_TABS[1]]);
             return redirect()->back()->withErrors([$e->getMessage()]);
         }
     }
